+++ packages/koot/CHANGELOG.md
@@ -27,6 +27,7 @@
     - **新** `internalLoaderOptions` - 用以扩展几乎无法修改的内置 `loader` 所用的设置。详情请参见文档的 [配置/Webpack](https://koot.js.org/#/config?id=internalloaderoptions) 章节
     - **新** `server.onRender.beforeDataToStore` 和 `server.onRender.afterDataToStore` - 允许更详细的使用服务器端渲染生命周期。详情请参见文档的 [配置/Webpack](https://koot.js.org/#/config?id=Webpack) 章节
     - `redux.syncCookie` 现支持传入 `true`: 同步所有 cookie，包括 cookie 原始字符串 (以 `__` 为名称)
+  - 优化 `koot-start` 命令，尽量避免 `koot-build 命令未找到` 的问题
   - Webpack 打包
     - 现在打包时不再会在项目根目录下生成临时文件
       - 这些文件现在移至 `/logs/tmp/` 目录下
@@ -78,6 +79,7 @@
     - `pm2` -> _3.2.4_
     - `portfinder` -> _1.0.20_
     - `postcss` -> _7.0.7_
+    - `rimraf` -> _2.6.3_
     - `webpack-dev-server` -> _3.1.14_
 
 ## 0.7.12